requirements.txt
aiogram==3.12.0
aiohttp==3.10.11
asyncpg==0.29.0
python-dotenv==1.2.1
APScheduler==3.11.1
httpx==0.28.1
pydantic==2.8.2